// src/frontend/src/types/currency.ts
/* 
 * This program has been developed by students from the bachelor Computer Science at 
 * Utrecht University within the Software Project course.
 * © Copyright Utrecht University (Department of Information and Computing Sciences)
 * Licensed under the MIT License. See the LICENSE file in the project root for details.
 */

export interface Currency {
    id: number;
    name: string;
    sprite: string;
}

export interface UserCurrency {
    userId: number;
    currencyId: number;
    amount: number;
}

export type UserCurrencyUpdateDto = Omit<UserCurrency, "userId">;